Fixed jitter and moved scroll to hooks file
